separated util classes and mixins
