+++ local_modules/styles/mixins.module.less
@@ -1,6 +1,6 @@
 @import "./variables.less";
 
-.container {
+.container () {
     margin-right: auto;
     margin-left: auto;
     padding-right: 15px;
@@ -9,7 +9,7 @@
     box-sizing: border-box;
 }
 
-.base-reset {
+.base-reset () {
     margin: 0;
     padding: 0;
     border: none;
@@ -17,46 +17,4 @@
     font: inherit;
     -webkit-appearance: none;
     appearance: none;
-}
-
-.italic-font {
-    color: @colorGrey;
-    font-style: italic;
-}
-
-.sorting-button {
-    &:extend(.italic-font, .base-reset);
-    cursor: pointer;
-}
-
-.loader {
-    display: flex;
-    flex-direction: column;
-    flex-grow: 1;
-    height: 100%;
-}
-
-.empty-result {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    color: @colorGrey;
-    height: 100vh - 50px;
-}
-
-.dropdown-checkbox {
-    display: flex;
-    align-items: center;
-    padding: 0;
-}
-
-.dropdown-checkbox-item {
-    padding: 6px 8px;
-}
-
-.noisinessSpinner {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    height: calc(100vh - 335px);
 }
