+++ local_modules/styles/utils.module.less
@@ -1,0 +1,45 @@
+@import "./variables.less";
+@import "./mixins.module.less";
+
+.italic-font {
+    color: @colorGrey;
+    font-style: italic;
+}
+
+.sorting-button {
+    .italic-font();
+    .base-reset();
+    cursor: pointer;
+}
+
+.loader {
+    display: flex;
+    flex-direction: column;
+    flex-grow: 1;
+    height: 100%;
+}
+
+.empty-result {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    color: @colorGrey;
+    height: 100vh - 50px;
+}
+
+.dropdown-checkbox {
+    display: flex;
+    align-items: center;
+    padding: 0;
+}
+
+.dropdown-checkbox-item {
+    padding: 6px 8px;
+}
+
+.noisinessSpinner {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    height: calc(100vh - 335px);
+}
