+++ src/Components/AllContatcsTable/AllContactsTable.module.less
@@ -1,3 +1,5 @@
+@import "~styles/utils.module.less";
+
 .cell {
     overflow: hidden;
     text-overflow: ellipsis;
