+++ src/Components/ContactEventStats/ContactEventStats.module.less
@@ -1,3 +1,5 @@
+@import "~styles/utils.module.less";
+
 .time-range-container {
     display: flex;
     justify-content: space-between;
