+++ src/Components/FilterStatusSelect/FilterStatusSelect.module.less
@@ -1,3 +1,5 @@
+@import "~styles/utils.module.less";
+
 .filter-status-icon:not(:last-child) {
     margin-right: 5px;
 }
