+++ src/Components/Header/Components/AdminMenu.module.less
@@ -1,5 +1,7 @@
+@import "~styles/utils.module.less";
+
 .admin-button {
-    &:extend(.base-reset);
+    .base-reset();
     color: #fff;
     cursor: pointer;
 }
