+++ src/Components/ManageSubscriptionsSelect/ManageSubscriptionsSelect.module.less
@@ -1,3 +1,5 @@
+@import "~styles/utils.module.less";
+
 .manage-btn {
     overflow: hidden;
     text-overflow: ellipsis;
