+++ src/Components/NotificationList/NotificationList.module.less
@@ -1,3 +1,5 @@
+@import "~styles/utils.module.less";
+
 .row {
     display: flex;
     align-items: flex-start;
