+++ src/Components/PatternList/PatternList.module.less
@@ -1,3 +1,5 @@
+@import "~styles/utils.module.less";
+
 .row {
     display: flex;
     flex-wrap: wrap;
@@ -20,7 +22,7 @@
     width: calc(~"100% - 240px");
     word-wrap: break-word;
     &.clicked {
-        &:extend(.base-reset);
+        .base-reset();
         text-align: left;
         cursor: pointer;
     }
