+++ src/Components/SubscriptionList/SubscriptionList.module.less
@@ -1,3 +1,5 @@
+@import "~styles/utils.module.less";
+
 .items-container {
     margin: 0 -10px;
 }
