+++ src/Components/TriggerInfo/TriggerInfo.module.less
@@ -16,9 +16,6 @@
     align-items: baseline;
     margin-left: auto;
 }
-.control {
-    margin-left: 15px;
-}
 .list {
     display: flex;
     flex-wrap: wrap;
